restructured into core/ input/ and ui/
