+++ src/core/cursor.rs
