+++ src/core/mod.rs
@@ -1,0 +1,9 @@
+
+
+pub mod app;
+pub mod editor;
+pub mod popups;
+pub mod cursor;
+pub mod command_line;
+pub mod errors;
+pub mod clipboard;
