+++ src/core/popups/confirmation_popup.rs
@@ -1,4 +1,4 @@
-use crate::popups::popup::{Popup, PopupResult, PopupType};
+use super::popup::{Popup, PopupResult, PopupType};
 use crossterm::event::KeyEvent;
 use ratatui::layout::{Alignment, Rect};
 use ratatui::prelude::{Color, Style};
