+++ src/core/popups/error_popup.rs
@@ -1,10 +1,10 @@
-use crate::popups::popup::{Popup, PopupResult, PopupType};
+use super::popup::{Popup, PopupResult, PopupType};
 use crossterm::event::KeyEvent;
 use ratatui::layout::{Alignment, Rect};
 use ratatui::prelude::{Color, Line, Span, Style, Text};
 use ratatui::widgets::{Block, Borders, Clear, Paragraph};
 use ratatui::Frame;
-use crate::errors::AppError;
+use super::super::errors::AppError;
 
 pub struct ErrorPopup {
     pub message: String,
