+++ src/lib.rs
@@ -1,12 +1,8 @@
 //expose modules
-pub mod app;
-pub mod editor;
-pub mod popups;
-pub mod cursor;
+pub mod core;
+
 pub mod config;
-pub mod command_line;
-pub mod clipboard;
-pub mod errors;
+
 pub mod ui;
 pub mod input;
